chore(intl): split Temporal toLocaleString into a sibling module (file-size gate)
The icu4x CLDR date-time pattern work grew date_collator.rs to 2182 lines,
past the 2000-line CI cap (scripts/check_file_size.sh). Move the
`TemporalLocaleCtx` enum + `temporal_locale_string` (the shared Temporal
`toLocaleString` implementation, 256 lines) into a child module
`date_collator/temporal.rs`, re-exported `pub(crate) use temporal::{...}`.
As a child module, `use super::*` still reaches the parent's private
formatting helpers (format_components / format_date_style / format_time_style
/ opt_string), so no call sites change and intl.rs is untouched. Trunk is
now 1936 lines; no logic moved.
